+++ AssessementProjectForAddingUser.Application/Interface/IServices/IAddingUserService.cs
@@ -7,7 +7,7 @@ namespace AssessementProjectForAddingUser.Application.Interface.IServices
         Task<ResponseDto> AddingUserInDb(UserDetailsAnkitDtos userDetailsAnkitDtos);
         Task<bool> LoginCredentialChecking(LoginCredentialDto loginCredential);
         Task<ResponseDto> GetAllUsers();
-        Task<ResponseDto> DeleteUserDetail(int Id);
+        Task<ResponseDto> DeleteUserDetail(long Id);
         Task<ResponseDto> UpdateUserDetail(UserDetailsAnkitDtos detailsAnkitDtos);
         Task<ResponseDto> SendEmailToForgotPassword(string email);
